Eighth commit

+++ libfractal/fractal.h
@@ -7,7 +7,7 @@ typedef struct fractal {
   int width; //uint32_t ?
   int height; //uint32_t ?
   int ** value;
-  double mean;
+  long double mean;
   double a, b;
 } fractal_t;
